Rename hooks to match existing filename convention
