+++ src/popup/components/toast_area/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { useToast } from '../../hooks/useToast';
+import { useToast } from '../../hooks/use_toast';
 
 import './toast_area.css';
